+++ src/org/lukasj/idea/torquescript/asset/AssetImporter.kt
@@ -9,7 +9,7 @@ import com.intellij.openapi.progress.util.ProgressWindow
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.util.io.exists
+import kotlin.io.path.exists
 import kotlinx.coroutines.flow.consumeAsFlow
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeout